requirements.txt
asyncio==3.4.3
h2==2.6.2
PyDictionary==1.5.2
wikipedia
psycopg2==2.7.4
youtube_dl==2018.11.07
pynacl==1.2.1
discord.py[voice] ~= 0.16.0
pip
colorlog
cffi --only-binary all; 
aiohttp>=1.0.0,<1.1.0
websockets>=3.1,<4.0
chardet
opuslib
BeautifulSoup4
rethinkdb
requests==2.20.0
ruamel.yaml
pendulum
Pillow==3.4.1
readline
psutil